+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to semantic versioning.
 
 - Add `dropRepeats` function to the `Array` module.
 - Add `dropLeft`, `dropLeftWhile`, and `dropRight` functions to the `String` module.
+- Add `empty` constant to the `String` module.
 - Add `unJSONString` function to the `JSON` module.
 - Remove `concat` function from the `String` module as it's a duplicate of `prepend`.